requirements.txt
torch
torchvision
torchmetrics
torchsummary
torchinfo
opencv-python
numpy
pyyaml
tqdm
wandb
tabulate